+++ src/string_utils.rs
@@ -16,7 +16,7 @@ pub fn is_line_ending(text: &str) -> bool {
     }
 }
 
-pub fn newline_count(text: &str) -> uint {
+pub fn newline_count(text: &str) -> usize {
     let mut count = 0;
     for c in text.chars() {
         if c == '\n' {
@@ -26,7 +26,7 @@ pub fn newline_count(text: &str) -> uint {
     return count;
 }
 
-pub fn char_count(text: &str) -> uint {
+pub fn char_count(text: &str) -> usize {
     let mut count = 0;
     for _ in text.chars() {
         count += 1;
@@ -34,7 +34,7 @@ pub fn char_count(text: &str) -> uint {
     return count;
 }
 
-pub fn grapheme_count(text: &str) -> uint {
+pub fn grapheme_count(text: &str) -> usize {
     let mut count = 0;
     for _ in text.graphemes(true) {
         count += 1;
@@ -42,7 +42,7 @@ pub fn grapheme_count(text: &str) -> uint {
     return count;
 }
 
-pub fn char_and_newline_count(text: &str) -> (uint, uint) {
+pub fn char_and_newline_count(text: &str) -> (usize, usize) {
     let mut char_count = 0;
     let mut newline_count = 0;
     
@@ -56,8 +56,8 @@ pub fn char_and_newline_count(text: &str) -> (uint, uint) {
     return (char_count, newline_count);
 }
 
-pub fn char_pos_to_byte_pos(text: &str, pos: uint) -> uint {
-    let mut i: uint = 0;
+pub fn char_pos_to_byte_pos(text: &str, pos: usize) -> usize {
+    let mut i: usize = 0;
     
     for (offset, _) in text.char_indices() {
         if i == pos {
@@ -73,8 +73,8 @@ pub fn char_pos_to_byte_pos(text: &str, pos: uint) -> uint {
     panic!("char_pos_to_byte_pos(): char position off the end of the string.");
 }
 
-pub fn grapheme_pos_to_byte_pos(text: &str, pos: uint) -> uint {
-    let mut i: uint = 0;
+pub fn grapheme_pos_to_byte_pos(text: &str, pos: usize) -> usize {
+    let mut i: usize = 0;
     
     for (offset, _) in text.grapheme_indices(true) {
         if i == pos {
